Rename credential variable names to current naming notation (#397)
⚠️ This is a breaking change ⚠️

This PR is small, but a breaking change that aims to keep the Ruby SDK consistent with the rest of the SDKs and the Nylas platform as a whole. Currently the Ruby SDK refers to the credentials as app_id and app_secret which Nylas has moved on from a long time ago. They have been changed to client_id and client_secret respectively to keep consistency with the platform.